FF-2853 fix: rename eppo_rb to eppo_client (#19)
* chore: fix Cargo.lock checking in CI

`cargo generate-lockfile` tries to update dependencies to latest
version and thus fails with `--locked`, even if the lockfile itself is
fine.

`cargo update --workspace` only updates files from the workspace root
and any packages that are missing in the lockfile. So it produces
consistent result.

* FF-2853 fix: rename eppo_rb to eppo_client

For some reason, ExtensionTask ignored lib_dir during gem
installation, so extension was installed to lib/eppo_rb and gem failed
to function.

Renaming import does not help because it breaks the build in local
development.

Rename eppo_rb to eppo_client, so extension name matches the
installation directory. This is also what was recommended by rust gem
template.