chore(ruby): fix ruby publish workflow (#27)
Our publish workflow uses Ruby 3.0 to push the gem. Ruby 3.0 ships
with RubyGems 3.2.3 which is below our minimum required
version (3.3.11).

The version used for publishing does not affect the result of
publishing. It's just using the old/unmaintained version of Ruby to do
the build and push itself (which also has security concerns).

Update Ruby in publish workflow to 3.3.