refactor(all),fix(rust): introduce Evaluator and handle SdkMetadata
Previously, eppo_core did not set metadata on events correctly (sdk
name and version were missing). Python and rust accounted for this
separately, and Rust didn't handle it at all.

This PR makes evaluation function take SdkMetadata as a parameter, so
it's now impossible to miss.

`now` is not also a parameter, so all evaluation functions are now
pure functions without side effects, so we can add time-based tests in
the future.

This PR also introduces Evaluator to simplify calling into evaluator
functions for upstream SDKs.